% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsid.R
\name{capsid}
\alias{capsid}
\title{capsid}
\usage{
capsid(
  h = 1,
  k = 0,
  H = NULL,
  K = NULL,
  t = capsid_turns,
  L = capsid_tilings,
  a = capsid_axes,
  R = 100,
  s = 0,
  roll = -90,
  pitch = 0,
  yaw = 0,
  penton_fiber = F,
  fiber_knob = F,
  facet_outline = F,
  aesthetics = list(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
description
}
