% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{facets}
\alias{facets}
\title{facets}
\usage{
facets(
  h = 1,
  k = 0,
  H = NULL,
  K = NULL,
  t = capsid_turns,
  L = capsid_tilings,
  a = capsid_axes,
  R = 100,
  s = 0,
  facet_outline = F,
  aesthetics = list(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
description
}
