% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsid.R, R/facets.R, R/lattice.R, R/net.R,
%   R/rdemocapsid.R
\name{capsidOutput}
\alias{capsidOutput}
\alias{renderCapsid}
\alias{facetsOutput}
\alias{renderFacets}
\alias{latticeOutput}
\alias{renderLattice}
\alias{netOutput}
\alias{renderNet}
\alias{rdemocapsid-shiny}
\title{Methods for use within the R Shiny framework.}
\usage{
capsidOutput(outputId, width = "100\%", height = "400px")

renderCapsid(expr, env = parent.frame(), quoted = FALSE)

facetsOutput(outputId, width = "100\%", height = "400px")

renderFacets(expr, env = parent.frame(), quoted = FALSE)

latticeOutput(outputId, width = "100\%", height = "400px")

renderLattice(expr, env = parent.frame(), quoted = FALSE)

netOutput(outputId, width = "100\%", height = "400px")

renderNet(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates a networkD3 graph}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Methods for use within the R Shiny framework.
}
