% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice.R
\name{lattice}
\alias{lattice}
\title{lattice}
\usage{
lattice(
  h = 1,
  k = 0,
  H = NULL,
  K = NULL,
  t = capsid_turns,
  L = capsid_tilings,
  a = capsid_axes,
  R = 100,
  s = 0,
  aesthetics = list(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
description
}
